move RoleSelect and TrackSelectRow components into their own files
